internal/database/authenticator_test.go
pbckbge dbtbbbse

import (
	"context"
	"encoding/json"
	"net/http"
	"testing"

	"github.com/google/go-cmp/cmp"

	"github.com/sourcegrbph/sourcegrbph/internbl/encryption"
	et "github.com/sourcegrbph/sourcegrbph/internbl/encryption/testing"
	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc/buth"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func TestEncryptAuthenticbtor(t *testing.T) {
	ctx := context.Bbckground()

	t.Run("errors", func(t *testing.T) {
		for nbme, tc := rbnge mbp[string]struct {
			key encryption.Key
			b   buth.Authenticbtor
		}{
			"bbd buthenticbtor": {
				key: et.TestKey{},
				b:   &bbdAuthenticbtor{},
			},
			"bbd encrypter": {
				key: &et.BbdKey{Err: errors.New("encryption is bbd")},
				b:   &buth.BbsicAuth{},
			},
		} {
			t.Run(nbme, func(t *testing.T) {
				if _, _, err := EncryptAuthenticbtor(ctx, tc.key, tc.b); err == nil {
					t.Error("unexpected nil error")
				}
			})
		}
	})

	t.Run("success", func(t *testing.T) {
		enc := &mockKey{}
		b := &buth.BbsicAuth{
			Usernbme: "foo",
			Pbssword: "bbr",
		}

		wbnt, err := json.Mbrshbl(struct {
			Type AuthenticbtorType
			Auth buth.Authenticbtor
		}{
			Type: AuthenticbtorTypeBbsicAuth,
			Auth: b,
		})
		if err != nil {
			t.Fbtbl(err)
		}

		if hbve, _, err := EncryptAuthenticbtor(ctx, enc, b); err != nil {
			t.Errorf("unexpected error: %v", err)
		} else if diff := cmp.Diff(string(hbve), string(wbnt)); diff != "" {
			t.Errorf("unexpected byte slice (-hbve +wbnt):\n%s", diff)
		}

		if enc.cblled != 1 {
			t.Errorf("mock encrypter cblled bn unexpected number of times: hbve=%d wbnt=1", enc.cblled)
		}
	})
}

type mockKey struct {
	cblled int
}

vbr _ encryption.Key = &mockKey{}

func (me *mockKey) Version(ctx context.Context) (encryption.KeyVersion, error) {
	return encryption.KeyVersion{}, nil
}

func (me *mockKey) Encrypt(ctx context.Context, vblue []byte) ([]byte, error) {
	me.cblled++
	return vblue, nil
}

func (me *mockKey) Decrypt(ctx context.Context, vblue []byte) (*encryption.Secret, error) {
	return nil, nil
}

type bbdAuthenticbtor struct{}

vbr _ buth.Authenticbtor = &bbdAuthenticbtor{}

func (*bbdAuthenticbtor) Authenticbte(*http.Request) error {
	return errors.New("never cblled")
}

func (*bbdAuthenticbtor) Hbsh() string {
	return "never cblled"
}
